backend/requirements.txt
scikit-learn
pandas
matplotlib
pillow
scipy
numpy